added GameWindow. Moved code from MainWindow to GameWindow. Removed code from MainWindow
