+++ arcadeGame/arcadeGame/GameWindow.xaml
@@ -1,5 +1,5 @@
-﻿<Window x:Class="arcadeGame.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+﻿<Window x:Class="arcadeGame.GameWindow"
+       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
@@ -7,7 +7,7 @@
         mc:Ignorable="d"
         Title="Space Invaders" Height="450" Width="800">
 
-    <Canvas Name="Canvas" Focusable="True" Background="Pink">
+    <Canvas Name="Canvas" Focusable="True" Background="Black" KeyDown="KeyPressed" KeyUp="KeyReleased">
         <Rectangle Name="player1" 
             Fill="White" 
             Height="66" 
@@ -34,6 +34,4 @@
 
         </Label>-->
     </Canvas>
-
-
 </Window>
